test(core-api): extract attachment and worker suites
